cambios 2

+++ app/src/main/java/net/gabrielf/apppedidos/EliminarActivity.java
@@ -6,12 +6,45 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.EditText;
+import android.widget.Toast;
 
-public class EliminarActivity extends AppCompatActivity {
+import com.mobsandgeeks.saripaar.ValidationError;
+import com.mobsandgeeks.saripaar.Validator;
+import com.mobsandgeeks.saripaar.annotation.NotEmpty;
+
+import java.util.List;
+
+public class EliminarActivity extends AppCompatActivity implements Validator.ValidationListener {
 
     DatabaseHelper dbHandler;
+    @NotEmpty(message = "Debe ingresar id a eliminar" )
     EditText elimina_input;
+    Validator validator;
 
+    @Override
+    public void onValidationSucceeded() {
+        //Toast.makeText(this, "Dato ingresado correctamente", Toast.LENGTH_SHORT).show();
+    }
+
+
+
+    @Override
+    public void onValidationFailed(List<ValidationError> errors)
+    {
+        for (ValidationError error : errors)
+        {
+            View view = error.getView();
+            String message = error.getCollatedErrorMessage(this);
+
+            if (view instanceof EditText) {
+                ((EditText) view).setError(message);
+            }
+            else
+            {
+                Toast.makeText(this, message, Toast.LENGTH_LONG).show();
+            }
+        }
+    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -19,13 +52,25 @@ public class EliminarActivity extends AppCompatActivity {
         setContentView(R.layout.activity_eliminar);
         elimina_input = (EditText) findViewById(R.id.modificar_input);
         dbHandler = new DatabaseHelper(this);
+
+        validator = new Validator(this);
+        validator.setValidationListener(this);
     }
 
     public void eliminar_clicked(View view){
 
-        dbHandler.borrarItems(Integer.parseInt(elimina_input.getText().toString()));
-        confirmacion();
+        EditText a = (EditText)findViewById(R.id.modificar_input);
+        String delete = a.getText().toString();
 
+        if (delete.equals("") || delete.equals(null)){
+
+            validator.validate();
+        }
+        else{
+            dbHandler.borrarItems(Integer.parseInt(elimina_input.getText().toString()));
+            confirmacion();
+            limpiarcampos();
+        }
 
     }
 
@@ -42,5 +87,11 @@ public class EliminarActivity extends AppCompatActivity {
                 });
         dlgAlert.setCancelable(true);
         dlgAlert.create().show();
+    }
+
+    public void limpiarcampos(){
+
+        elimina_input.setText("");
+
     }
 }
